+++ dmr/routing.py
@@ -14,6 +14,7 @@ from dmr.errors import ErrorType, format_error
 from dmr.exceptions import InternalServerError, NotAcceptableError
 from dmr.internal.routing import RouterMetadata
 from dmr.internal.routing import URLExternal as _URLExternal
+from dmr.internal.types import FormatError, StrOrPromise
 from dmr.openapi.collector import (
     collect_normalized_paths,
     controller_mapping_collector,
@@ -22,11 +23,6 @@ from dmr.openapi.objects import PathItem, Paths
 from dmr.openapi.openapi import OpenAPI
 
 if TYPE_CHECKING:
-    from django.utils.functional import (
-        _StrOrPromise,  # pyright: ignore[reportPrivateUsage]
-    )
-
-    from dmr.internal.types import FormatError
     from dmr.openapi.core.context import OpenAPIContext
     from dmr.renderers import Renderer
     from dmr.serializer import BaseSerializer
@@ -232,7 +228,7 @@ class Router:
 
 
 def external_path(
-    route: '_StrOrPromise',
+    route: StrOrPromise,
     view: _DjangoView,
     *,
     openapi: PathItem | None,
@@ -276,7 +272,7 @@ def build_404_handler(
     /,
     *prefixes: str,
     serializer: type['BaseSerializer'],
-    format_error: 'FormatError' = format_error,
+    format_error: FormatError = format_error,
     renderers: Sequence['Renderer'] | None = None,
 ) -> Callable[[HttpRequest, Exception], HttpResponse]:
     """
@@ -359,7 +355,7 @@ def build_500_handler(
     /,
     *prefixes: str,
     serializer: type['BaseSerializer'],
-    format_error: 'FormatError' = format_error,
+    format_error: FormatError = format_error,
     renderers: Sequence['Renderer'] | None = None,
 ) -> Callable[[HttpRequest], HttpResponse]:
     """
@@ -468,21 +464,21 @@ class _PrefixRoutePattern(RoutePattern):
 # NOTE: keep in sync with `django-stubs`!
 @overload
 def path(
-    route: '_StrOrPromise',
+    route: StrOrPromise,
     view: _DjangoView,
     kwargs: dict[str, Any] | None = None,
     name: str | None = None,
 ) -> URLPattern: ...
 @overload
 def path(
-    route: '_StrOrPromise',
+    route: StrOrPromise,
     view: tuple[Sequence[_AnyPattern], str | None, str | None],
     kwargs: dict[str, Any] | None = None,
     name: str | None = None,
 ) -> URLResolver: ...
 @overload
 def path(
-    route: '_StrOrPromise',
+    route: StrOrPromise,
     view: Sequence[URLResolver | str],
     kwargs: dict[str, Any] | None = None,
     name: str | None = None,
@@ -490,7 +486,7 @@ def path(
 
 
 def path(
-    route: '_StrOrPromise',
+    route: StrOrPromise,
     view: (
         _DjangoView
         | tuple[Sequence[_AnyPattern], str | None, str | None]